docs(Toggle): Moved toggle over to new configurable example
