+++ src/components/form/example/index.js
@@ -5,6 +5,5 @@ module.exports = [
   require('./text-inputs-themes').default,
   require('./text-area').default,
   require('./check-boxes').default,
-  require('./toggle').default,
   require('./radio-buttons').default,
 ];
